fix(Todo): refactorizacion de las funciones del orm

See merge request novakelloggs/todo-api!6

+++ src/apps/todo/models/todo.py
@@ -18,12 +18,35 @@ class Todo(db.Model):
         db.session.commit()
         return self
 
-    # FIXME: Validar si se obtiene un objeto
     def delete(self):
-        db.session.delete(self)
-        db.session.commit()
-        return self
+        if self != None:
+            db.session.delete(self)
+            db.session.commit()
+            return self
+        return {}
 
     def to_dict(self):
         return {c.key: getattr(self, c.key)
                 for c in inspect(self).mapper.column_attrs}
+
+    @classmethod
+    def fetch_all(cls):
+        todos = cls.query.all()
+        todo_list = []
+
+        for todo in todos:
+            todo_list.append(todo.to_dict())
+
+        return todo_list
+
+    @classmethod
+    def find_by_id(cls, id):
+        return cls.query.get(id).to_dict() if cls.query.get(id) else {}
+
+    @classmethod
+    def create(cls, name):
+        return cls(name).save().to_dict() if name != None else {}
+
+    @classmethod
+    def remove_by_id(cls, id):
+        return cls.query.get(id).delete().to_dict() if cls.query.get(id) else {}
